production_process backend 중간 merge

+++ monmi/src/main/java/com/monmi/domain/Material.java
@@ -14,12 +14,12 @@ public class Material {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="material_id")
-    private Long materialId;
+    private int materialId;
     @Column(name="material_name", nullable = false)
     private String materialName;
 
     @Builder
-    public Material(Long materialId, String materialName) {
+    public Material(int materialId, String materialName) {
         this.materialId = materialId;
         this.materialName = materialName;
     }
